requirements.txt
aiohttp==3.6.2
aiosqlite==0.13.0
alembic==1.4.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
iso8601==0.1.12
Mako==1.1.3
MarkupSafe==1.1.1
multidict==4.7.6
PyPika==0.37.7
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
tortoise-orm==0.16.13
typing-extensions==3.7.4.2
yarl==1.4.2
